Add tool to set service on/off from maintenance
Turning the service into maintenance mode is a tasks that requires
multiple map/unmap routes commands.

To avoid the requirement to have to do this manually each time we
want to set the service on or off from maintenance, with the possibility
of executing the wrong command or the confusion for new developers in
the service, we have written a Thor client tool that automates this
process.

This tool allows as to set the production pages on/off from maintenance,
and will ensure that the required services are available before
attempting to do re-mappings.