Rely in "only downgrade" flag over specific states
The "only downgrade" flag is being used in a singular case.
Is that case where, on top of the flag, we're hardcoding the particular
transition of states we want it to be checked over.

The change makes the code not to depend on particular states. If the flag is
set, we will only allow state downgrades.