Remove need/usage of DISABLE_OVERRIDES_FOR
Noticed that the DISABLE_OVERRIDES_FOR usage is needed because we were
storing the current state as previous state prior to check if we need
to override the update by using the previous state.

The DISABLE_OVERRIDES_FOR was ensuring that the previous and current
states are not matching (we are already in what is set as previous)
before attempting the override.

To resolve this, we store the current step in a temp variable and we do
not attempt to update the previous_state until the state has been
changed.

This way the logic depending on the previous state can be simplified.