Do not create Nanomaterials until name is set
When adding a nanomaterial to the notification draft, we don't want the
nanomaterial to be created until the nanomaterial name has been
introduced.

If an user clicks on "Add" in the nanomaterials draft section, will be
redirected to the "Add a nanomaterial INCI name" page. From there they
will be able to click on the backlink and be redirected to the draft
page without having a new nanomaterial in the nanomaterials list.

If, the user sets the name instead, and then clicks on the back link
twice to be sent back to the draft page, they will find the name of the
new nanomaterial listed in the draft page.