Remove unreachable branch from nanom count update
The view won't show the contains nanomaterials form unless the
notification has no nanomaterials at all.

The logic depending on that form update, was also considering the
possibility of notifications already having ONE nanomaterial before the
update.

As this possibility is unreachable for the user, we are removing it from
the back-end logic.