Do not create nanomaterials or items from notification draft until introducing their names (#2451)
* Remove "build" dir from gitignore file

As is so generic that causes "build" wizard folders to be ignored.

* Do not create Nanomaterials until name is set

When adding a nanomaterial to the notification draft, we don't want the
nanomaterial to be created until the nanomaterial name has been
introduced.

If an user clicks on "Add" in the nanomaterials draft section, will be
redirected to the "Add a nanomaterial INCI name" page. From there they
will be able to click on the backlink and be redirected to the draft
page without having a new nanomaterial in the nanomaterials list.

If, the user sets the name instead, and then clicks on the back link
twice to be sent back to the draft page, they will find the name of the
new nanomaterial listed in the draft page.

* Refactor: Remove unused duplicated view pages

These pages are living inside the wizard/nanomaterial/build folder now.

* Refactor: Remove deprecated helper methods

- One is pointing to a deleted route.

- Other method has been replaced with the new flow back routing
  constants and wizard concern method to map them.

* Refactor: Remove views left from old flow

Not used by new flow

* Do not create Components until their name is set

As done with the nanomaterials, when adding a component/item from the
notification draft page, the component won't be added/listed in the
draft until the form for setting the component name is submitted.

* Refactor: Move controllers into Notifications namespace

Product and Product kit controllers belong to notifications namespace.
Renamed and wrapped them inside Notifications module/folder for
domain coherence.

* Refactor: Remove unused resource/routes

Do not match any controller, as it was moved/renamed with the new
workflow.

* Remove Wizard namespace from controllers/forms

Reducing the resource nesting to the minimum necessary to match the
business domain.

* Remove developer comments

* Remove build folder from cfignore file

(cherry picked from commit 622da6851e375a3cc3f7d77d83d2255241f1d29f)