Re-design pages for  nanos/items are present.
When a notification draft already has nanomaterials or items present and
the user re-runs the product creation task, it will be displayed a page
indicating to add/remove the nano/item from the draft instead of the
original form to introduce a number of nanos/items to be created.

This work re-designs those pages.