Fix sign-in flaky feature tests
We had these tests randomnly failing for a while on our automated test
runs, forcing us to re-run the specs in order to unblock the Pull
Request for merging.

Cause of the flaky specs:
In order to calculate a wrong OTP Code, we were reversing the user OTP
Code generated by the service while having 2FA enabled.

This is easy and works well when the generated code is "12345", having
"54321" as wrong code.
But what happens when the user OTP code is a palyndrome? "12121" turns
into... "12121". So the "wrong code" is the same as the right code. And
when asserting over entering the wrong code and getting an error... the
test fails because it actually managed to sign-in with the right code.

Resolved it by replacing the first digit of the OTP code. If it is a '0'
it will be come '1'. If not, will decrease the value by 1.

This way we ensure we will still have 5 digits and the resulting code
will be different, even with palyndromes.