Refactor: Wrong service error pages
Refactor of the error pages displayed when a Search user attempts to
visit a Submit service page or a Submit user attempts to visit a Search
service page.

We renamed templates to be closer to our service domain language.
So instead of "business account" we use "submit user".
Also unified both views for poison_centres/msa into a single "search",
as both pages were sharing 99% of their content besides some wording
based on the user role.