Fix RP invitation to new unconfirmed user emai
When a Responsible Person team member invitation was sent to an email
address that matches an existing user "new email" that has not been
confirmed, accepting the invitation was causing the creation of a new
user.

The new user have the email address of the existing user new email.
This causes multiple issues:
- Both users become invalid due to sharing email values.
- Both users in our service belong to a single person.

To resolve this, when an user receives a team invitation to their new
unconfirmed email, and follows the link inside the valid confirmation
time period:

1. No new user will be created. Instead will find the user by new_email
   and add this user to the team.
2. Since the user opened this confirmation link from their new_email
   inbox, this invitation email will also confirm the email change. So,
   on top of joining the team, the user will have their email change to
   the new_email address.