Fix label images display for MSA users (#2498)
The view for label images is shared between MSA users and Submit users.

When the label image has not passed the antivirus check, it shows a
"Refresh" link that reloads the edit page for Submit users.

The same link was being called from Search MSA users, causing a routing
error (as MSA users are not associated with a Responsible Person and the
route parameter is missing).

MSA users shouldn't even know if the image is processing or not. Just
having it displayed when is available and not displayed when it is not.

The fix only shows the link if is edits are allowed for the user when
loading the page (submit user belonging to the RP who created the
notification).