Add ingredients exact concentration new notification (#2501)
* Add new page HTML for exact concentration

* Extract CAS number validations to custom validator

* Add CAS number and Poisonous check to Exact Formula

Cas number is optional and follows a specific format.

Poisonous check is a boolean value to identify if an ingredient is
poisonous or not.

* Fix checkbox padding

* Do not allow empty names/quantities on exact formulas

Since these are the key values for an exact formula object. It
does not make sense to have any of them without those values.

Production DB does not contain empty names/quantities, so it is safe to
add this restriction.

* Basic adding exact ingredients flow

Allows users to add exact ingredients one after another.
Allows to move one when finished adding ingredients.
Validations over the ingredient exact concentration fields.
Validations over the form asking if the user wants to add more ingredients.

* Allow skipping exact concentration form

When a component has already an ingredient exact concentration, we can
skip the introduction of new ingredients exact concentration.
It will send us directly to the PH question step.

* Show a list of already added ingredients

On the page for adding ingredients exact concentrations.

* Validate unique ingredient names per component

When trying to introduce new ingredients in the exact concentration
form, the form won't allow users to introduce an ingredient name that
has been already added.

* Fix array style rule issue

* Use product/item wording based on single/multi-com

When the notification is single component, we will use "product" in the
wording. When the notification is multi-component, we will use "item".

* Add casting types to form attributes

So they cast the values from the form inputs.

* Cover form with unit tests

* Bring the latest CSS changes from the prototype project

* Adapt feature specs to new exact ingredients flow.

Feature specs that were adding exact ingredients have been adapted to
use the new page for adding ingredients instead of the PDF concentration
upload.

* Move Exact Formula creation to form object

Makes code simpler and is a common pattern for form objects.

* Validate positive value is given as concentration

Avoiding having values like "random" cast into a 0.0 as a float,
stored in Database and displayed.
We also want to avoid inputs as "0", as they need to introduce a
positive value as concentration.