Manually add ingredients as range concentrations
Instead of uploading the component ingredients as a PDF that contains
the ingredients concentration ranges, we are requiring to manually
introduce each ingredient and select their concentration range.

Users will be shown a flow loop where they can add one by one each of
the component ingredients.

When the ingredient is poisonous, the user must introduce the exact
concentration of the ingredient, and later can continue introducing
other poisonous/non poisonous ingredients.