New frame formulation poisonous ingredient flow
When user selects to add poisonous ingredients to a frame formulation
component will be shown a form to manually add each poisonous ingredient
and their exact concentration.

The form and flow will be almost identical to adding exact ingredients
to a component, but with slightly differences in the wording and with
the "Poisonous" check being pre-selected and enforced (users won't be
able to uncheck them).