Scenario 4: Delete ings on formulation type change
When an user is editing the component details and reach the formulation
type selection page, they will see a warning message on top of the
selection, letting them know that changing the selection will wipe any
existing ingredient.

As warned, when the user changes the formulation type, any existing
ingredient will be deleted, and the next page for the user will be a
blank "Add ingredient" page.