Allow to revisit and edit existing ingredients (#2529)
* Scenario 1: Editing ingredients already added

When users try to redo an Item details tasks from the draft, and the
task already had ingredients. They will be shown the already existing
ingredients one by one, allowing them to edit their values and types
(range component ingredient changed from poisonous to not poisonous or
the other way).

* Scenario 2: Adding an ingredient after editing.

After editing the previously existing ingredients, users will be sent to
the "Add another ingredient" page.
The difference will be that the page, after editing existing
ingredients, won't display the "The ingredient was successfully added"
banner message.

If answering yes, the form for adding new ingredients will show as
usual, including the "Skip" link (not available on edit pages). On
success adding a new ingredient, the "Add another ingredient" page will
contain the success banner this time.

* Scenario 3: Removing an ingredient

Users will be able to remove ingredients when re-visiting their edit
details page.

Depending on the ingredient that is removed, the user will either be sent
to:
- Component formulation type selection (if it was the only ingredient
  for the component).
- Question for adding further new ingredients (if removed last
  ingredient for a component with multiple ingredients).
- Next ingredient edit page (if removing an ingredient but not the
  last).

* Scenario 4: Delete ings on formulation type change

When a user is editing the component details and reaches the formulation
type selection page, they will see a warning message on top of the
selection, letting them know that changing the selection will wipe any
existing ingredient.

As warned, when the user changes the formulation type, any existing
ingredient will be deleted, and the next page for the user will be a
blank "Add ingredient" page.

* Scenario 5: Do not allow to submit legacy drafts

Legacy drafts with formulation files instead of individual ingredients
won't be valid and users won't be able to confirm their submission.

When revisiting the formulation type wizard page, the current formulation
file-pdf attachments will be purged and the user will go through the current
flow of manually adding ingredients.

* Increase the allowed input length for exact input

* Stricter exact number validation

Raises a numericality validation error instead of truncating the number
when the user introduces an exact concentration with non-numeric
characters. Eg: "12.45:3" was being stored as "12.45". It will get a
validation error now.