Allows to edit frame formulations poisonous ing.
As donw with range/exact formulations, when a user revisits the
component details tasks, will find the existing ingredients (in this
case the poisonous ingredients) prefilled one by one. The user will be
able to edit the details or to remove the ingredients.

When changing the formulation type from predefined to exact/range all
the existing predefined formulation poisonous ingredients will be wiped.
Same if when revisiting the flow the user answers "No" to the "contains
poisonous ingredients?" question.