Skip DB ingredient name uniqueness on legacy data
There are many legacy notifications (imported via ZIP file) and few
deleted notifications that contain multiple ingredients with the same
name for the same component.

As we only allow uniquely named ingredients for a component, this would
become an issue when trying to migrate legacy data into the new
Ingredient model.

We are resolving it by excluding legacy/deleted notifications from the
unique name ingredient validation.