Fix states after formulation type change
When the user changes the formulation type in a completed component, the
existing component ingredients get removed and neither the component or
the notification should be displayed as "completed" (or ready to be
completed in the notification case) anymore.

The component state needs to be pushed back to "empty" from "completed"
and the notification to "ready for components" from "components
complete".

In this way, when the user changes the formulation type but does not
complete the component with ingredients, and then goes back to the
notification draft, the component will be displayed as empty instead of
complete, and the "Accept and Submit" link for the notification will be
disabled until the component is finally completed.