Fix: Display empty shade fields when none present
The check over the component having "nil" as shades value works fine for
pre-creating 2 empty shades when reaching the "add shades" page for the
first time.

The problem is that, when the user skips the shades addition by not
introducing any value in the shades fields, the component shades value
will become an empty array "[]".

This is not very consistent from the model data point, and probably we
will want to migrate all production data to ensure we go for either
"nil" or "[]" as shades attribute value when the component has no
shades.

Besides that, there is an user interface issue when not filling any
shades and then re-visiting the step moving back to the same point
again.

With the current code, a component with shades "[]" won't match the
condition to display to empty shades in the form, so the form will be
displayed without any input field.

Checking that there is no shades (empty array or nil) resolves this and
ensures user flow consistency. Either when visiting it for the first
time or revisiting it.