Fix error when removing images without JS-enabled
When browsing without JS enabled, the link to remove an image was
triggering a server error.
This is because the delete links map to an AJAX call that cannot be
executed without JavaScript.

Fixed it by using a deletion form with a button calling the deletion
with a POST request instead.