Adapt "Add Nanomaterial" from draft page
This flow had a hardcoded "Add name" first page that, upon completion,
was creating a Nanomaterial/Nanoelement and redirecting to the build
nanomaterial flow next step after name introduction.

Since the build nanomaterial flow has been reshaped to ask first for the
purposes and later for the name, we needed to also adapt the "Add
Nanomaterial" from draft page so it starts with the purposes form, and
then redirects to the "Add name" page in the build nanomaterial flow.