Change nanomaterial journey (#2587)
* Add link to anex in "add nanomaterial name" page

In the notification builder, when adding a nanomaterial, now the input
hint for the nanomaterial name has a link to the Annex 4 list of
colourants document.

* Refactor: Create NanoMaterialPurposes model

There was a lot of logic and constants around the Nano Material Purposes
concept.

Wrapped this concept into a NanoMaterialPurposes PORO, declared as a non
ActiveRecord model.

This simplifies the helpers logic and provides a single source of truth
for Nano Material Purposes across the service.

* Refactor nanomaterial helper methods naming

- Inline helper methods that is not used anywhere but by another helper.
- Rename helpers to remove "get_" prefixing.

* Refactor: Remove redundant helper method

Now that we have Purpose objects we can query attributes over them
without complex logic. This can be called directly when building the
checkbox items list.

* Move Nano Material purposes question before name

Swap the Nano Material wizard pages, setting the purposes selection as
the first step and the name as second step.

* Anidate purposes selection in wizard form

Instead of displaying all the options in a single checkbox multi-choice
selection, we display a radio selection with 2 options:
- Predefined purposes: Colourant, Preservative or UV filter.
- Another purpose.

When selecting the radio button for the predefined purposes, a
multi-choice checkbox will expand with all the predefined options.

* Refactor: Use method for non standard purpose

* Refactor: Replace strings with constants

* Use Form object for nano material purposes form

Allows us to:
- introduce specific validations in the purpose type selection and the
  purposes.
- pre-select form elements when displaying an existing selection from
  existing nano material in database.
- infer purposes from type selection, and type value from purposes
  selection.

* Rename "predefined" to "standard" purposes

To match the domain language for the nano materials.

* Reword validation errors

To match the wording specified in the ticket requirements.

* Fix display of annex link

* Adapt wizard steps routing to new flow

- Only request the name when selecting standard purposes for the
  nanomaterial.

* Fix controller specs to work with new flow

* Fix title errors

Show the error in the title if the form has any validation errors, not
only when errors belong to "purposes" attribute.

* Extract purposes form to its own partial

So it can be used by different views.

* Adapt "Add Nanomaterial" from draft page

This flow had a hardcoded "Add name" first page that, upon completion,
was creating a Nanomaterial and redirecting to the build
nanomaterial flow next step after name introduction.

Since the build nanomaterial flow has been reshaped to ask first for the
purposes and later for the name, we needed to also adapt the "Add
Nanomaterial" from draft page so it starts with the purposes form, and
then redirects to the "Add name" page in the build nanomaterial flow.

* Rubocop fixes

* List all the selected purposes anexes in name page