Add task to ensure 1 nanoelement per nanomaterial
There are few cases of nanomaterials containing 2 nanoelements in
production db.
We want each nanoelement belonging to a single nanomaterial.

To resolve this, for each nanoelement on top of the first in the
nanomaterial, we will create a clone of the current nanomaterial,
associate it to the notification and component to match the current
nanomaterial associations and, finally, associate the nanoelement to the
cloned nanomaterial.