Add nanomaterials rake maintenance tasks (#2593)
* Add task to delete nanomaterials without nanoelem.

Any nanomaterial that has no associated nanoelement is invalid and
should be deleted from our DB.

Even when they belong to a submitted notification, we only display
nanoelement names and info in the notification info.
It is safe to delete nanomaterials that have no nanoelement and
consequently no name.

* Add task to delete orphan nanomaterials

Between Jan and October 2021 28.5k nanomaterials were created without
being associated with any notification or component.

This seems to be caused by some bug/error when importing zip
notifications. Either caused by the importing task or by the imported
data itself.

Whatever the reason, these nanomaterials are not associated with any
notification, so they should be deleted together with any nanoelement
they may contain.

* Add task to ensure 1 nanoelement per nanomaterial

There are a few cases of nanomaterials containing 2 nanoelements in
production DB.
We want each nanoelement to belong to a single nanomaterial.

To resolve this, for each nanoelement on top of the first in the
nanomaterial, we will create a clone of the current nanomaterial,
associate it to the notification and component to match the current
nanomaterial associations and, finally, associate the nanoelement to the
cloned nanomaterial.