Remove redundant migrations on Formulas tables
Originally we develop the new ingredient flow using the existing
ExactFormula and RangeFormula models.
In order to achieve the desired flow, we had to add a few fields in
those tables.

Since we ended creating a new Ingredient model, the changes over
Exact/RangeFormula models can be reverted, as these changes never went
live and no data was added in the new fields.