Fix: Nano keeping inci_name after purpose changes
When a nanomaterial is standard, the nanomaterial has an inci_name, but
when the nanomaterial is not standard (other purposes), the name will
come from the associated nanomaterial notification.

There was a bug where, when originally filling standard nanomaterial and
introducing a name for it, the user could go back and then change the
purposes to "other", complete the non-standard purposes route, but still
being shown the original standard nanomaterial name.

Fixing it by removing the inci_name when the user select "other"
purposes.