Do not allow to introduce long ingredient names (#2674)
We are introducing a form validation to not allow very long text
introduced as an ingredient name when creating an ingredient.

Since we moved from PDF formulations to requiring RP users to introduce
their product ingredients manually, some users used the lack of length
restriction on the name to paste the whole product formulation
(sometimes thousands of character) as an ingredient.

This validation will avoid that for future ingredients.
What to do with the current ingredients/products violating this will be
decided in the future. At the moment those product notifications should
pass validations, hence not introducing a model validation toghether
with the form one.