Fix follow team invite link with incomplete user
ISSUE

When:
1. A team invited a non registered user email.
2. The user registered themselves without following the link.
3. User "sign out" from account security page instead of completing
   their registration.
4. User follows up the invitation link to register/join the team.

Then:
A sign-in page with a "Your login credentials were used in another
browser." error. Instead of displaying the account security page and
allowing the user to complete their registration while getting added to
the Responsible Person that invited them.

CAUSE

The method used for signing in an user from an invitation link, was
calling "bypass_sign_in" Devise method. That method bypasses Warden
callbacks and stores the user straight in session.
Thr problem with that, is that one of those Warden callbacks sets
the

warden.user.submit_user.session"=>{"unique_session_id"=>"value"}

in the user session. This value is used by Devise SessionLimitable
module to check what is the current session value against the unique
session id stored in DB for the current user.

As it wasn't set on the session, but the user was marked as "signed in"
in their session, the SessionLimitable module takes this as the user
current session not matching the unique session id in DB, and triggers
the "Your login credentials were used in another browser" error.

SOLUTION

Do a default Devise sign_in, including Warden callbacks,  when following
the team invitation link. This sets the unique session id from the user
DB into the session: "unique_session_id"=>"value".

With this set, following the link passes the SessionLimitable check and
the user ends in the account security form to complete their
registration, as expected.