Refactor: Rename Ingredients sort_by as group_by
The current Ingredient search filter is displayed as "Group by" in the
user interface, even when internally is performing a results sorting.

As we are introducing a differest "Sort by" filter near the search box,
we are going to differenciate both by renaming the existing one to match
the interface ("Group by"), while the new one will be introduced as
"Sort by".