Fix missing params in backlinks from notifications
The parameters used to search notifications are also present in the
poison centre notifications controller in order to generate backlinks
from the notification view page to the exact search that triggered the
result.

Added a note to avoid future issues, as both the grouping and sorting
params were missing from ingredients search.