Remove filter of OpenSearch results state
We were filtering out non completed notifications from the OpenSearch
results as, between the period of a notification being deleted and the
next daily full re-index, empty notifications (deleted from the service
but not removed from the index), were being returned in the search
results.

As we are now deleting them individually from the index at the same time
the notification gets deleted from the service, this "completed" check
is no longer necesary when displaying the index search results.