Remove individual deleted notification from Opensearch index (#2692)
* Refactor: Reduce spec nesting

By extracting aliased method specs outside of method they are aliased
from.

* Refactor: Use described_class in specs

* Rely on state machine for setting "deleted" state

Instead of manually setting the "delete" state and the "deleted_at"
timestamp from the notification#soft_delete! call, we are using the
notification state machine to trigger a transition and automatically set
the deletion timestamp for the notification.

* Remove notification from search index when deleted

When soft deleting a notification, the notification will be removed from
the open search index, and will not be available in the search results.

Followed the pattern of adding the notification to the index from a
state machine callback, doing the same to remove the notification.

* Remove destroyed notification from search index

As done with the soft delete, when a hard delete completely removes a
notification record from DB we also want to remove this notification
from the opensearch index.

* Remove filter of OpenSearch results state

We were filtering out non completed notifications from the OpenSearch
results as, between the period of a notification being deleted and the
next daily full re-index, empty notifications (deleted from the service
but not removed from the index), were being returned in the search
results.

As we are now deleting them individually from the index at the same time
the notification gets deleted from the service, this "completed" check
is no longer necesary when displaying the index search results.