Bugfix: Notifications missing deletion timestamps (#2699)
* Remove notification deletion logic from the state machine

The change of the state to "deleted" through the state machine
transition was not writing the deleted_at timestamp, even when it was
supposed to do so.

It seems this was caused of the change of state "without validations"
restricting the update to the state column.

This was passing a test specifically written to check the deletion
timestamp being written. Why when passing the test wasn't writing the
timestamps in production... no idea.

Moved all this logic (and the deletion from the index) to the original
method containing all the soft deletion logic in the model.
I don't trust the state machine at this point :P.

* Add task to backfill missing deleted_at timestamps

As the updated_at contains the last date the notification got modified
(when deleted), the value for deleted_at can be set to updated_at.

We are avoiding updating the updated_at timestamp when backfilling the
deleted_at column.