Changes on Cosmetic Search date filters for users navigating without JavaScript (#2696)
* Add hide class for non js to ing. search reset

* Add hide class for non js to product filter radios

* Remove unused ID locale from partial calls

The "id" parameters passed to the partial in every date filter input
call is not used. The inputs id is derived by the attribute parameter.

* Allow date filter input to have custom IDs

Now we will allow the date filter input to use custom ID naming if it is
specified when rendering the partial.
When not specified, will default to the attribute value.

* Add no JS date filters to product search

It will allow search users to filter de notification dates when
searching products without using a browser with JavaScript.

* Remove filters section padding

* Hide Reset from non-js notifications date filter

As done for the ingredients. Reset link needs JS to remove input current
values. So for non-js users it won't work.