Move PG advisory lock to a module
We are creating a module that will allow us to conditionally execute
code inside a Postgres transaction advisory lock.

The code to execute will be a block, and the module method requires a
string to identify the particular lock to attempt.

In this way, we can execute "only if not already running" transactions
across our service.