Allow to hot-swap ES indexes through ES aliases
At the moment, the service is periodically doing a full index import but
not re-creating the index. This caused some deleted notifications being
stored in the index and not deleted.

Apparently the reasons for not recreating the index was that sometimes
the import would fail and the index was already deleted/created again,
having no results in the search service as a consequence.

This would also be an issue once the import turn long, as the live
service would start from an empty search index that slowly populates.
Returned data would be incomplete during reindexing times.

We want to achieve re-creating elasticsearch indexes from scratch
without having service downtime.

To achieve that, we are using elasticsearch aliases, that allow us to
point the indexed model to an alias name, while the alias name points to
a current index.
Then we will be able to re-create a new index from scratch and hot-swap
it in the alias. So now the alias will point to the new index.

All with 0 downtime for the users.