Implement zero downtime full OpenSearch Reindex
Now that we can hot-swap indices on a running app through the use of
elasticsearch aliases, we are re-implementing the full Opensearch
reindexing (deleting an index and building it from scratch) without
having service downtime.

The process is based on creating a new index, importing all the
qualifying notifications to the new index, checking if the import was
successfull and swapping the actual index used by the application
through re-pointing the alias to it.
Then the previous index gets deleted.

This will allow users to keep getting search results in the service
while the whole reindexing happens in the background. And once the
alias gets pointed to the new index this will inmediately available to
further searches.