Add a scheduled job to delete old indices versions
Sometimes, while reindexing, the job may be killed (EG: during a
re-deploy) once the new index was already created but before the import
finished and the alias was pointed to it.

This index won't get deleted in further re-indexings as it has never
been aliased to the model alias.
This task clean up possible indices left behind.