Reintroduce opensearch reindexing (#2706)
* Move PG advisory lock to a module

We are creating a module that will allow us to conditionally execute
code inside a Postgres transaction advisory lock.

The code to execute will be a block, and the module method requires a
string to identify the particular lock to attempt.

In this way, we can execute "only if not already running" transactions
across our service.

* Allow to hot-swap ES indexes through ES aliases

At the moment, the service is periodically doing a full index import but
not re-creating the index. This caused some deleted notifications being
stored in the index and not deleted.

Apparently the reasons for not recreating the index was that sometimes
the import would fail and the index was already deleted/created again,
having no results in the search service as a consequence.

This would also be an issue once the import turn long, as the live
service would start from an empty search index that slowly populates.
Returned data would be incomplete during reindexing times.

We want to achieve re-creating elasticsearch indexes from scratch
without having service downtime.

To achieve that, we are using elasticsearch aliases, that allow us to
point the indexed model to an alias name, while the alias name points to
a current index.
Then we will be able to re-create a new index from scratch and hot-swap
it in the alias. So now the alias will point to the new index.

All with 0 downtime for the users.

* Remove no longer relevant TODO comment

* Implement zero downtime full OpenSearch Reindex

Now that we can hot-swap indices on a running app through the use of
elasticsearch aliases, we are re-implementing the full Opensearch
reindexing (deleting an index and building it from scratch) without
having service downtime.

The process is based on creating a new index, importing all the
qualifying notifications to the new index, checking if the import was
successfull and swapping the actual index used by the application
through re-pointing the alias to it.
Then the previous index gets deleted.

This will allow users to keep getting search results in the service
while the whole reindexing happens in the background. And once the
alias gets pointed to the new index this will inmediately available to
further searches.

* Remove eager loading of lib folder

* Add start of the reindexing info to logs

* Increase Review APPs worker memory from 500M to 2GB

* Covers first time setting alias on reindex

When the alias does not exist and we attemted to swap it, ES was
triggering an exception.
Cover this case by creating the alias in case it doesn't exist.

* Remove redundant job name from log calls

The job name is already included by Sidekiq logger, so was redundant.

* Ensure previous index exists before deletion

* Fix permissions on delete review app index script

The script didn´t have execution permissions.

* Reduce OpenSearch re-indexing frequencey to 1/week

Since we are adding/removing individual documents to the elasticsearch
index, the re-indexing to sync any wrong document from the index can be
done less frequently than once per day.

Is an expensive task so the less we run it, the better.

* Follow Rails conventions for Concern class methods

* Reduce method naming verbosity

* Rename reindex job scheduling task in sidekiq

* Add a scheduled job to delete old indices versions

Sometimes, while reindexing, the job may be killed (EG: during a
re-deploy) once the new index was already created but before the import
finished and the alias was pointed to it.

This index won't get deleted in further re-indexings as it has never
been aliased to the model alias.
This task clean up possible indices left behind.

* Rubocop fix

* Remove unused job

There is no code invoking this job.

* Add README info on Opensearch management

Include the usual add/remove process and the particular jobs for
re-indexing and deleting dangling indeces, together with reasoning
behind the process.

* Tweaks over postgres lock

- Rename Postgres lock class.
- Extract lock name to a constant.

* Move index/delete document methods to Searchable

Indexing and deleting from index are methods that belong to the
searching domain.

* Add logging to individual ElasticSearch operations

Each operation over elasticsearch will have its own individual log
entry.

* Wrap indices deletion with logging

So every time indices are deleted there will be a log entry.

* Rename "old/previous" with "unused"

To be more descriptive with the actual behaviour, as unused indices
newer than the "current" one used by the model will also be identified
and deleted.

* Allow to pass index to ".import_to_opensearch"

If an index is provided to the method, use that index for the import.

* Use ".import_to_opensearch" when reindexing

Use the wrapped method instead of directly calling
ES library.

* Ensure index log tag is shared across the service

Reindex job and the individual index actions will be tagged under the
same log tag for easier tracing through the logs.

Co-authored-by: Marc Sardón <marc@bichito.software>

* Fix specs

Co-authored-by: Marc Sardón <marc@bichito.software>

* Restore namespaced log method name on shared code

As this is part of a concern and will become part of the public
interface on any class using this concern.

Co-authored-by: Sławosz Sławiński <slawosz@gmail.com>